Add HoverStateMixin.
