+++ patterns/state/manipulator_state/states/creations/creation_state.dart
@@ -2,11 +2,12 @@ import 'dart:ui';
 
 import '../../pattern/manipulator_context.dart';
 import '../../shapes/shape.dart';
+import '../hover_state_mixin.dart';
 import '../selections/selection_state.dart';
 import '../specific_actions_mixin.dart';
 
 abstract class CreationState extends ManipulationState
-    with SpecificActionsMixin {
+    with SpecificActionsMixin, HoverStateMixin {
   var _startX = 0.0;
   var _startY = 0.0;
   Shape? _newShape;
