+++ patterns/state/manipulator_state/states/free_sate.dart
@@ -1,7 +1,9 @@
 import '../pattern/manipulator_context.dart';
+import 'hover_state_mixin.dart';
 import 'specific_actions_mixin.dart';
 
-class FreeState extends ManipulationState with SpecificActionsMixin {
+class FreeState extends ManipulationState
+    with SpecificActionsMixin, HoverStateMixin {
   @override
   void mouseDown(double x, double y) {
     trySelectShape(x, y);
